refactor(#38): picker 전체 component로 이동
