+++ app/src/main/java/com/ssafy/neegongnaegong/presentation/calendar/detail/ScheduleDetailScreen.kt
@@ -34,7 +34,7 @@ import com.ssafy.neegongnaegong.domain.model.calendar.RepeatType
 import com.ssafy.neegongnaegong.domain.model.calendar.Schedule
 import com.ssafy.neegongnaegong.presentation.calendar.component.CalendarTopAppBar
 import com.ssafy.neegongnaegong.presentation.calendar.component.ScheduleEditText
-import com.ssafy.neegongnaegong.presentation.calendar.component.picker.datetime.DateTimeRangePicker
+import com.ssafy.neegongnaegong.presentation.component.picker.datetime.DateTimeRangePicker
 import com.ssafy.neegongnaegong.presentation.component.LoadingDialog
 import com.ssafy.neegongnaegong.presentation.ui.theme.NeeGongNaeGongTheme
 import kotlinx.coroutines.flow.Flow
