+++ app/src/main/java/com/ssafy/neegongnaegong/presentation/calendar/edit/ScheduleEditScreen.kt
@@ -40,7 +40,7 @@ import com.ssafy.neegongnaegong.domain.model.calendar.UpdateType
 import com.ssafy.neegongnaegong.presentation.calendar.component.CalendarTopAppBar
 import com.ssafy.neegongnaegong.presentation.calendar.component.RepeatRuleInput
 import com.ssafy.neegongnaegong.presentation.calendar.component.ScheduleEditText
-import com.ssafy.neegongnaegong.presentation.calendar.component.picker.datetime.DateTimeRangePicker
+import com.ssafy.neegongnaegong.presentation.component.picker.datetime.DateTimeRangePicker
 import com.ssafy.neegongnaegong.presentation.component.LoadingDialog
 import com.ssafy.neegongnaegong.presentation.ui.theme.NeeGongNaeGongTheme
 import kotlinx.coroutines.flow.Flow
