+++ app/src/main/java/com/ssafy/neegongnaegong/presentation/component/picker/datetime/DateTimeRangePicker.kt
@@ -1,4 +1,4 @@
-package com.ssafy.neegongnaegong.presentation.calendar.component.picker.datetime
+package com.ssafy.neegongnaegong.presentation.component.picker.datetime
 
 import androidx.compose.animation.AnimatedVisibility
 import androidx.compose.animation.Crossfade
@@ -28,8 +28,8 @@ import androidx.compose.ui.draw.scale
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.ssafy.neegongnaegong.presentation.calendar.component.ScheduleEditText
-import com.ssafy.neegongnaegong.presentation.calendar.component.picker.time.TimePicker
-import com.ssafy.neegongnaegong.presentation.calendar.component.picker.date.DateRangePicker
+import com.ssafy.neegongnaegong.presentation.component.picker.time.TimePicker
+import com.ssafy.neegongnaegong.presentation.component.picker.date.DateRangePicker
 import com.ssafy.neegongnaegong.presentation.ui.theme.NeeGongNaeGongTheme
 import java.time.LocalDateTime
 import java.time.YearMonth
